Add Embeddings behavior
