+++ lib/sanbase/knowledge/faq.ex
@@ -53,7 +53,7 @@ defmodule Sanbase.Knowledge.Faq do
 
   def find_similar_entries(user_input, size \\ 5) do
     with {:ok, [user_embedding]} <-
-           Sanbase.OpenAI.Embedding.generate_embedding([user_input], 1536) do
+           Sanbase.AI.Embedding.generate_embeddings([user_input], 1536) do
       query =
         from(
           e in FaqEntry,
@@ -158,7 +158,7 @@ defmodule Sanbase.Knowledge.Faq do
     Answer: #{entry.answer_markdown}
     """
 
-    case Sanbase.OpenAI.Embedding.generate_embedding([text], 1536) do
+    case Sanbase.AI.Embedding.generate_embeddings([text], 1536) do
       {:ok, [embedding]} ->
         entry
         |> Ecto.Changeset.change(embedding: embedding)
